Fix rounding of scaled page size
With the previous implementation, fitting the page width to the
screen width sometimes resulted in a page width that was actually
one pixel *larger* than the screen width. This was due the use
of Math.round(). This commit should fix this issue by always
rounding _down_ the would-be floating point page dimensions to
the integer page dimensions.